requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.10
gunicorn==22.0.0
requests==2.31.0